+++ src/index.js
@@ -1,65 +1,55 @@
 /* @flow */
-/* eslint-disable react/no-find-dom-node */
 
 import fromEventsWithOptions from './lib/fromEventsWithOptions';
 import Kefir from 'kefir';
 import kefirBus from 'kefir-bus';
 import type {Bus} from 'kefir-bus';
 import React from 'react';
-import type {Element as ReactElement} from 'react';
-import {createPortal, findDOMNode} from 'react-dom';
+import {createPortal} from 'react-dom';
 import PropTypes from 'prop-types';
 import containByScreen from 'contain-by-screen';
 import type {Options} from 'contain-by-screen';
 import isEqual from 'lodash/isEqual';
 
-const requestAnimationFrame = global.requestAnimationFrame || (cb => setTimeout(cb, 0));
+const requestAnimationFrame = global.requestAnimationFrame || (cb => Promise.resolve().then(cb));
 
-type FloatAnchorContext = {
-  repositionEvents: Object;
+type FloatAnchorContextType = {
+  repositionEvents: Kefir.Observable<null>;
 };
+
+// Context is used so that when a FloatAnchor has reposition() called on it,
+// all of its descendant FloatAnchor elements reposition too.
+const FloatAnchorContext = React.createContext<?FloatAnchorContextType>(null);
 
 export type {Options} from 'contain-by-screen';
 
 export type Props = {
-  anchor: ReactElement<any>;
-  float?: ?ReactElement<any>;
+  anchor: (anchorRef: React$Ref<any>) => React$Node;
+  float?: ?React$Node;
   options?: ?Options;
   zIndex?: ?number|string;
   floatContainerClassName?: ?string;
 };
 export default class FloatAnchor extends React.Component<Props> {
   static propTypes = {
-    anchor: PropTypes.element.isRequired,
-    float: PropTypes.element,
+    anchor: PropTypes.func.isRequired,
+    float: PropTypes.node,
     options: PropTypes.object,
     zIndex: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
     floatContainerClassName: PropTypes.string
   };
 
+  static contextType = FloatAnchorContext;
+
   _portalEl: ?HTMLElement;
   _portalRemoval: Bus<null> = kefirBus();
   _unmount: Bus<null> = kefirBus();
   _repositionEvents: Bus<null> = kefirBus();
-
-  // Context is used so that when a FloatAnchor has reposition() called on it,
-  // all of its descendant FloatAnchor elements reposition too.
-  static childContextTypes = {
-    floatanchor: PropTypes.object
+  _childContext: FloatAnchorContextType = {
+    repositionEvents: this._repositionEvents
   };
-  static contextTypes = {
-    floatanchor: PropTypes.object
-  };
-  getChildContext(): Object {
-    const floatanchor: FloatAnchorContext = {
-      repositionEvents: this._repositionEvents
-    };
-    return {floatanchor};
-  }
 
-  _parentCtx(): ?FloatAnchorContext {
-    return this.context.floatanchor;
-  }
+  _anchorRef: ?HTMLElement = null;
 
   static *parentNodes(node: Node): Iterator<Node> {
     do {
@@ -77,9 +67,22 @@ export default class FloatAnchor extends React.Component<Props> {
     }
   }
 
+  _setAnchorRef = (anchorRef: ?HTMLElement) => {
+    this._anchorRef = anchorRef;
+
+    const portalEl = this._portalEl;
+    if (portalEl) {
+      if (anchorRef) {
+        (portalEl: any).rfaAnchor = anchorRef;
+      } else {
+        delete (portalEl: any).rfaAnchor;
+      }
+    }
+  };
+
   componentDidMount() {
     this._updateFloat();
-    const parentCtx = this._parentCtx();
+    const parentCtx: ?FloatAnchorContextType = this.context;
     if (parentCtx) {
       parentCtx.repositionEvents
         .takeUntilBy(this._unmount)
@@ -93,10 +96,6 @@ export default class FloatAnchor extends React.Component<Props> {
   }
 
   componentDidUpdate(prevProps: Props) {
-    if (this._portalEl && prevProps.anchor !== this.props.anchor) {
-      (this._portalEl: any).rfaAnchor = findDOMNode(this);
-    }
-
     if (this._portalEl && prevProps.floatContainerClassName !== this.props.floatContainerClassName) {
       this._portalEl.className = this.props.floatContainerClassName || '';
     }
@@ -120,21 +119,21 @@ export default class FloatAnchor extends React.Component<Props> {
 
   _updateFloat() {
     const {float} = this.props;
+    const portalEl = this._portalEl;
 
     if (float) {
-      if (!this._portalEl) throw new Error('Should not happen: portalEl not initialized');
-      const portalEl = this._portalEl;
+      if (!portalEl) throw new Error('Should not happen: portalEl not initialized');
 
       portalEl.style.zIndex = String(this.props.zIndex);
       portalEl.style.position = 'fixed';
 
       if (!portalEl.parentElement) {
-        const el = findDOMNode(this);
-        if (!el) throw new Error('ReactFloatAnchor missing element');
+        const anchorRef = this._anchorRef;
+        if (!anchorRef) throw new Error('ReactFloatAnchor missing anchorRef element');
         const target = document.body || document.documentElement;
         if (!target) throw new Error('Could not find element to attach portal to');
         target.appendChild(portalEl);
-        (portalEl: any).rfaAnchor = el;
+        (portalEl: any).rfaAnchor = anchorRef;
         this._portalRemoval.take(1).onValue(() => {
           (portalEl: any).rfaAnchor = undefined;
           if (portalEl.parentElement) portalEl.parentElement.removeChild(portalEl);
@@ -142,7 +141,7 @@ export default class FloatAnchor extends React.Component<Props> {
         Kefir.merge([
           Kefir.fromEvents(window, 'resize'),
           fromEventsWithOptions(window, 'scroll', {capture: true, passive: true})
-            .filter(event => event.target.contains(el))
+            .filter(event => event.target.contains(anchorRef))
         ])
           .takeUntilBy(this._portalRemoval)
           .onValue(() => {
@@ -150,7 +149,7 @@ export default class FloatAnchor extends React.Component<Props> {
           });
       }
     } else {
-      if (this._portalEl && this._portalEl.parentElement) {
+      if (portalEl && portalEl.parentElement) {
         this._portalRemoval.emit(null);
       }
     }
@@ -158,10 +157,9 @@ export default class FloatAnchor extends React.Component<Props> {
 
   reposition() {
     const portalEl = this._portalEl;
-    if (portalEl && portalEl.parentElement) {
-      const el = findDOMNode(this);
-      if (!(el instanceof HTMLElement)) throw new Error('ReactFloatAnchor missing element');
-      containByScreen(portalEl, el, this.props.options || {});
+    const anchorRef = this._anchorRef;
+    if (portalEl && portalEl.parentElement && anchorRef) {
+      containByScreen(portalEl, anchorRef, this.props.options || {});
       this._repositionEvents.emit(null);
     }
   }
@@ -169,23 +167,23 @@ export default class FloatAnchor extends React.Component<Props> {
   render() {
     const {anchor, float} = this.props;
     let floatPortal = null;
-    if (float) {
-      if (!this._portalEl) {
-        this._portalEl = document.createElement('div');
+    if (float != null) {
+      let portalEl = this._portalEl;
+      if (!portalEl) {
+        portalEl = this._portalEl = document.createElement('div');
       }
-
-      const portalEl = this._portalEl;
-      floatPortal = createPortal(float, portalEl);
+      floatPortal = (
+        <FloatAnchorContext.Provider value={this._childContext}>
+          {createPortal(float, portalEl)}
+        </FloatAnchorContext.Provider>
+      );
     }
 
-    // Using this small trick instead of an array so anchor and floatPortal
-    // don't need keys. TODO Use <>...</> instead or whatever official fragment
-    // technique React adds in the future.
-    return ((
-      <div>
-        {anchor}
+    return (
+      <>
+        {anchor((this._setAnchorRef: any))}
         {floatPortal}
-      </div>
-    ): any).props.children;
+      </>
+    );
   }
 }